add:hook_followbutton
