+++ frontend/src/components/atoms/FollowButton.tsx
@@ -1,43 +1,9 @@
-import { FC, useState, useEffect } from 'react';
+import { FC } from 'react';
 import { Icon, Button } from 'semantic-ui-react';
-import { useParams, useHistory } from 'react-router-dom';
-import { FetchFollow, FetchFollowed, FetchFollowExists } from 'apis/Follow';
+import useFollowButton from 'hooks/FollowButton';
 
 const FollowButton: FC = () => {
-  const [state, setState] = useState(false);
-  const { id } = useParams<{ id: string }>();
-  const history = useHistory();
-
-  useEffect(() => {
-    FetchFollowExists(id)
-      .then((result) => result === 200 && setState(() => true))
-      .catch(() =>
-        history.push('/', { message: 'エラーが発生しました。', type: 'error' }),
-      );
-  }, [id, history]);
-
-  const onFollow = () =>
-    FetchFollow(id)
-      .then((result) =>
-        result === 201
-          ? setState((prev) => !prev)
-          : history.push(`/registration/${id}`, {
-              message: 'エラーが発生しました。',
-              type: 'error',
-            }),
-      )
-      .catch();
-  const onFollowed = () =>
-    FetchFollowed(id)
-      .then((result) =>
-        result === 201
-          ? setState((prev) => !prev)
-          : history.push(`/registration/${id}`, {
-              message: 'エラーが発生しました。',
-              type: 'error',
-            }),
-      )
-      .catch();
+  const { state, onFollow, onFollowed } = useFollowButton();
 
   return state ? (
     <Button
